+++ smath/src/index.ts
@@ -300,6 +300,67 @@ export abstract class SMath {
         return Math.sqrt(this.vars(data));
     }
     /**
+     * Generate a uniformly-distributed floating-point number within the range.
+     * @param min The minimum bound
+     * @param max The maximum bound
+     * @returns A random float within the range
+     * @example
+     * ```js
+     * const y = SMath.runif(-2, 2); // 0.376...
+     * ```
+     */
+    public static runif(min: number, max: number): number {
+        return this.expand(Math.random(), min, max);
+    }
+    /**
+     * Generate a uniformly-distributed integer within the range.
+     * @param min The minimum bound (inclusive)
+     * @param max The maximum bound (inclusive)
+     * @returns A random integer within the range
+     * @example
+     * ```js
+     * const y = SMath.rint(-4, 3); // -4
+     * ```
+     */
+    public static rint(min: number, max: number): number {
+        min |= 0;
+        max |= 0;
+        if (min < 0) { min--; }
+        if (max < 0) { max--; }
+        return this.runif(min, max + 1) | 0; // `| 0` pulls toward 0
+    }
+    /**
+     * Generate a normally-distributed floating-point number.
+     * @param mean The mean of the population distribution
+     * @param stdev The standard deviation of the population
+     * @returns A random float
+     * @example
+     * ```js
+     * const y = SMath.rnorm(2, 3); // 1.627...
+     * ```
+     */
+    public static rnorm(mean: number = 0, stdev: number = 1): number {
+        return mean + stdev * Math.sqrt(-2 * Math.log(Math.random())) * Math.cos(2 * Math.PI * Math.random());
+    }
+    /**
+     * Generate a population of normally-distributed floating-point numbers.
+     * @param count The number of values to generate
+     * @param mean The mean of the population distribution
+     * @param stdev The standard deviation of the population
+     * @returns A population of random floats
+     * @example
+     * ```js
+     * const dataset = SMath.rdist(3); // [ 1.051..., -0.779..., -2.254... ]
+     * ```
+     */
+    public static rdist(count: number, mean: number = 0, stdev: number = 1): Array<number> {
+        const distribution: Array<number> = [];
+        for (let i = 0; i < count; i++) {
+            distribution[i] = this.rnorm(mean, stdev);
+        }
+        return distribution;
+    }
+    /**
      * Take the limit of a function. A return value of `NaN` indicates
      * that no limit exists either due to a discontinuity or imaginary value.
      * @param f Function `f(x)`